Organize classes into Duke package
Classes exist individually.

By grouping them under the same package, we can easily identify the relevant classes to the project.
