+++ src/main/java/Duke/Task.java
@@ -1,3 +1,5 @@
+package Duke;
+
 import java.util.ArrayList;
 
 public class Task {
