+++ src/main/java/Duke/Todo.java
@@ -1,3 +1,5 @@
+package Duke;
+
 import java.util.ArrayList;
 
 public class Todo extends Task {
